fix(arm32): widen MinidumpContextARM::fpscr to uint64_t
MinidumpContextARM::fpscr was declared as uint32_t, making the struct
364 bytes on disk. Every other downstream reader of the ARM32 minidump
context stream — breakpad's MDRawContextARM, rust-minidump's
CONTEXT_ARM, the Sentry ingest path that uses rust-minidump — treats
fpscr as uint64_t and expects a 368-byte struct, so crashpad's output
was 4 bytes short and couldn't be fully deserialized.

The comment above the struct already noted that extra[8] is
"included for compatibility with breakpad"; fpscr's width was the
remaining divergence. Switch it to uint64_t so sizeof(MinidumpContextARM)
== 368, matching breakpad / rust-minidump.

rust-minidump's struct at a pinned SHA, for reference:

  https://github.com/rust-minidump/rust-minidump/blob/d4fefc765aad35b3bef569d53c1680eadab5a268/minidump-common/src/format.rs#L1034-L1055

+++ minidump/minidump_context.h
@@ -355,6 +355,7 @@ enum MinidumpContextARMFlags : uint32_t {
 };
 
 //! \brief A 32-bit ARM CPU context (register state) carried in a minidump file.
+//! \see https://github.com/rust-minidump/rust-minidump/blob/d4fefc765aad35b3bef569d53c1680eadab5a268/minidump-common/src/format.rs#L1034-L1055
 struct MinidumpContextARM {
   //! \brief A bitfield composed of values of #MinidumpContextFlags and
   //!     #MinidumpContextARMFlags.
@@ -375,7 +376,7 @@ struct MinidumpContextARM {
   uint32_t cpsr;
 
   //! \brief Floating-point status and control register.
-  uint32_t fpscr;
+  uint64_t fpscr;
 
   //! \brief VFP registers `d0`-`d31`.
   uint64_t vfp[32];
